+++ src/Record.js
@@ -308,6 +308,32 @@ export default Component.extend({
   },
 
   /**
+   * Return whether the record is unsaved. Records that have primary keys are
+   * considered "saved". Records without primary keys are considered "unsaved".
+   *
+   * @example <caption>Record#isNew</caption>
+   * // Normally you would do: import {Container} from 'js-data'
+   * const JSData = require('js-data@3.0.0-beta.7')
+   * const {Container} = JSData
+   * console.log('Using JSData v' + JSData.version.full)
+   * const store = new Container()
+   * store.defineMapper('user')
+   * const user = store.createRecord('user', {
+   *   id: 1234
+   * })
+   * const user2 = store.createRecord('user')
+   * console.log('user isNew: ' + user.isNew()) // false
+   * console.log('user2 isNew: ' + user2.isNew()) // true
+   *
+   * @method Record#isNew
+   * @returns {boolean} Whether the record is unsaved.
+   * @since 3.0.0
+   */
+  isNew (opts) {
+    return utils.isUndefined(utils.get(this, this._mapper().idAttribute))
+  },
+
+  /**
    * Return whether the record in its current state passes validation.
    *
    * @example <caption>Record#isValid</caption>
@@ -580,8 +606,18 @@ export default Component.extend({
     const mapper = this._mapper()
     const id = utils.get(this, mapper.idAttribute)
     let props = this
+
+    const postProcess = (result) => {
+      const record = opts.raw ? result.data : result
+      if (record) {
+        utils.deepMixIn(this, record)
+        this.commit()
+      }
+      return result
+    }
+
     if (utils.isUndefined(id)) {
-      return superMethod(mapper, 'create')(props, opts)
+      return superMethod(mapper, 'create')(props, opts).then(postProcess)
     }
     if (opts.changesOnly) {
       const changes = this.changes(opts)
@@ -589,14 +625,7 @@ export default Component.extend({
       utils.fillIn(props, changes.added)
       utils.fillIn(props, changes.changed)
     }
-    return superMethod(mapper, 'update')(id, props, opts).then((result) => {
-      const record = opts.raw ? result.data : result
-      if (record) {
-        utils.deepMixIn(this, record)
-        this.commit()
-      }
-      return result
-    })
+    return superMethod(mapper, 'update')(id, props, opts).then(postProcess)
   },
 
   /**