+++ test/unit/mapper/create.test.js
@@ -131,38 +131,49 @@ describe('Mapper#create', function () {
       }
     }
 
-    let group = await store.create('group', store.createRecord('group', {
+    let group
+    group = await store.create('group', store.createRecord('group', {
       users: [{ name: 'John' }]
     }), { with: ['users'] })
     assert(group.users[0].id)
     assert.equal(group.users[0].name, 'John')
 
+    let user
     // when props are a Record
-    let user = await store.create('user', store.createRecord('user', JSData.utils.copy(userProps)), { with: [] })
+    user = await store.createRecord('user', JSData.utils.plainCopy(userProps)).save({ with: [] })
     assert(store.is('user', user), 'user should be a user record')
     assert.strictEqual(store.get('user', user.id), user, 'user should be in the store')
-    assert.deepEqual(user.comments, [], 'user.comments should be an empty array')
+    assert.objectsEqual(user.comments, [
+      {
+        content: 'foo',
+        userId: user.id
+      }
+    ], 'user.comments should be an array with a comment')
     assert.deepEqual(store.getAll('comment'), [], 'comments should not be in the store')
-    assert(!user.profile, 'user.profile should be undefined')
+    assert.objectsEqual(user.profile, {
+      email: 'john@email.com'
+    }, 'user.profile should be undefined')
     assert.deepEqual(store.getAll('profile'), [], 'profile should not be in the store')
     assert(!user.organization, 'user.organization should be undefined')
     assert(!user.organizationId, 'user.organizationId should be undefined')
     assert.deepEqual(store.getAll('organization'), [], 'organization should not be in the store')
     clear()
 
-    user = await store.create('user', store.createRecord('user', JSData.utils.copy(userProps)), { with: ['comment'] })
+    user = await store.createRecord('user', JSData.utils.plainCopy(userProps)).save({ with: ['comments'] })
     assert(store.is('user', user), 'user should be a user record')
     assert.strictEqual(store.get('user', user.id), user, 'user should be in the store')
     assert(store.is('comment', user.comments[0]), 'user.comments[0] should be a comment record')
     assert.objectsEqual(store.getAll('comment'), user.comments, 'comments should be in the store')
-    assert(!user.profile, 'user.profile should be undefined')
+    assert.objectsEqual(user.profile, {
+      email: 'john@email.com'
+    }, 'user.profile should be undefined')
     assert.deepEqual(store.getAll('profile'), [], 'profile should not be in the store')
     assert(!user.organization, 'user.organization should be undefined')
     assert(!user.organizationId, 'user.organizationId should be undefined')
     assert.deepEqual(store.getAll('organization'), [], 'organization should not be in the store')
     clear()
 
-    user = await store.create('user', store.createRecord('user', JSData.utils.copy(userProps)), { with: ['comment', 'profile'] })
+    user = await store.create('user', store.createRecord('user', JSData.utils.plainCopy(userProps)), { with: ['comments', 'profile'] })
     assert(store.is('user', user), 'user should be a user record')
     assert.strictEqual(store.get('user', user.id), user, 'user should be in the store')
     assert(store.is('comment', user.comments[0]), 'user.comments[0] should be a comment record')
@@ -174,7 +185,7 @@ describe('Mapper#create', function () {
     assert.deepEqual(store.getAll('organization'), [], 'organization should not be in the store')
     clear()
 
-    user = await store.create('user', store.createRecord('user', JSData.utils.copy(userProps)), { with: ['comment', 'profile', 'organization'] })
+    user = await store.create('user', store.createRecord('user', JSData.utils.plainCopy(userProps)), { with: ['comments', 'profile', 'organization'] })
     assert(store.is('user', user), 'user should be a user record')
     assert.strictEqual(store.get('user', user.id), user, 'user should be in the store')
     assert(store.is('comment', user.comments[0]), 'user.comments[0] should be a comment record')
@@ -187,43 +198,56 @@ describe('Mapper#create', function () {
     clear()
 
     // when props are NOT a record
-    user = await store.create('user', JSData.utils.copy(userProps), { with: [] })
+    user = await store.create('user', JSData.utils.plainCopy(userProps), { with: [] })
     assert(store.is('user', user), 'user should be a user record')
     assert.strictEqual(store.get('user', user.id), user, 'user should be in the store')
-    assert.deepEqual(user.comments, [], 'user.comments should be an empty array')
+    assert.objectsEqual(user.comments, [
+      {
+        content: 'foo',
+        userId: user.id
+      }
+    ], 'user.comments should be an array with a comment')
     assert.deepEqual(store.getAll('comment'), [], 'comments should not be in the store')
-    assert(!user.profile, 'user.profile should be undefined')
+    assert.objectsEqual(user.profile, {
+      email: 'john@email.com'
+    }, 'user.profile should not be undefined')
     assert.deepEqual(store.getAll('profile'), [], 'profile should not be in the store')
-    assert(!user.organization, 'user.organization should be undefined')
+    assert.objectsEqual(user.organization, {
+      name: 'Company Inc'
+    }, 'user.organization should not be undefined')
     assert(!user.organizationId, 'user.organizationId should be undefined')
     assert.deepEqual(store.getAll('organization'), [], 'organization should not be in the store')
     clear()
 
-    user = await store.create('user', JSData.utils.copy(userProps), { with: ['comment'] })
+    user = await store.create('user', JSData.utils.plainCopy(userProps), { with: ['comments'] })
     assert(store.is('user', user), 'user should be a user record')
     assert.strictEqual(store.get('user', user.id), user, 'user should be in the store')
     assert(store.is('comment', user.comments[0]), 'user.comments[0] should be a comment record')
     assert.objectsEqual(store.getAll('comment'), user.comments, 'comments should be in the store')
-    assert(!user.profile, 'user.profile should be undefined')
+    assert.objectsEqual(user.profile, {
+      email: 'john@email.com'
+    }, 'user.profile should not be undefined')
     assert.deepEqual(store.getAll('profile'), [], 'profile should not be in the store')
-    assert(!user.organization, 'user.organization should be undefined')
+    assert.objectsEqual(user.organization, {
+      name: 'Company Inc'
+    }, 'user.organization should not be undefined')
     assert(!user.organizationId, 'user.organizationId should be undefined')
     assert.deepEqual(store.getAll('organization'), [], 'organization should not be in the store')
     clear()
 
-    user = await store.create('user', JSData.utils.copy(userProps), { with: ['comment', 'profile'] })
+    user = await store.create('user', JSData.utils.plainCopy(userProps), { with: ['comments', 'profile'] })
     assert(store.is('user', user), 'user should be a user record')
     assert.strictEqual(store.get('user', user.id), user, 'user should be in the store')
     assert(store.is('comment', user.comments[0]), 'user.comments[0] should be a comment record')
     assert.objectsEqual(store.getAll('comment'), user.comments, 'comments should be in the store')
     assert(store.is('profile', user.profile), 'user.profile should be a profile record')
     assert.objectsEqual(store.getAll('profile'), [user.profile], 'profile should be in the store')
-    assert(!user.organization, 'user.organization should be undefined')
+    assert(store.is('organization', user.organization), 'user.organization should be a organization record')
     assert(!user.organizationId, 'user.organizationId should be undefined')
     assert.deepEqual(store.getAll('organization'), [], 'organization should not be in the store')
     clear()
 
-    user = await store.create('user', JSData.utils.copy(userProps), { with: ['comment', 'profile', 'organization'] })
+    user = await store.create('user', JSData.utils.plainCopy(userProps), { with: ['comments', 'profile', 'organization'] })
     assert(store.is('user', user), 'user should be a user record')
     assert.strictEqual(store.get('user', user.id), user, 'user should be in the store')
     assert(store.is('comment', user.comments[0]), 'user.comments[0] should be a comment record')
@@ -236,7 +260,7 @@ describe('Mapper#create', function () {
     clear()
 
     assert.equal(createCalledCount.user, 9)
-    assert.equal(createCalledCount.comment, 9)
+    assert.equal(createCalledCount.comment, 6)
     assert.equal(createCalledCount.profile, 4)
     assert.equal(createCalledCount.organization, 2)
   })
@@ -298,31 +322,40 @@ describe('Mapper#create', function () {
     }
 
     // when props are a Record
-    let user = await store.create('user', store.createRecord('user', JSData.utils.copy(userProps)), { pass: [] })
+    let user = await store.create('user', store.createRecord('user', JSData.utils.plainCopy(userProps)), { pass: [] })
     assert(store.is('user', user), 'user should be a user record')
     assert.strictEqual(store.get('user', user.id), user, 'user should be in the store')
-    assert.deepEqual(user.comments, [], 'user.comments should be an empty array')
+    assert.objectsEqual(user.comments, [
+      {
+        content: 'foo',
+        userId: user.id
+      }
+    ], 'user.comments should be an array')
     assert.deepEqual(store.getAll('comment'), [], 'comments should not be in the store')
-    assert(!user.profile, 'user.profile should be undefined')
+    assert.objectsEqual(user.profile, {
+      email: 'john@email.com'
+    }, 'user.profile should be a profile')
     assert.deepEqual(store.getAll('profile'), [], 'profile should not be in the store')
     assert(!user.organization, 'user.organization should be undefined')
     assert(!user.organizationId, 'user.organizationId should be undefined')
     assert.deepEqual(store.getAll('organization'), [], 'organization should not be in the store')
     clear()
 
-    user = await store.create('user', store.createRecord('user', JSData.utils.copy(userProps)), { pass: ['comment'] })
+    user = await store.create('user', store.createRecord('user', JSData.utils.plainCopy(userProps)), { pass: ['comments'] })
     assert(store.is('user', user), 'user should be a user record')
     assert.strictEqual(store.get('user', user.id), user, 'user should be in the store')
     assert(store.is('comment', user.comments[0]), 'user.comments[0] should be a comment record')
     assert.objectsEqual(store.getAll('comment'), user.comments, 'comments should be in the store')
-    assert(!user.profile, 'user.profile should be undefined')
+    assert.objectsEqual(user.profile, {
+      email: 'john@email.com'
+    }, 'user.profile should be a profile')
     assert.deepEqual(store.getAll('profile'), [], 'profile should not be in the store')
     assert(!user.organization, 'user.organization should be undefined')
     assert(!user.organizationId, 'user.organizationId should be undefined')
     assert.deepEqual(store.getAll('organization'), [], 'organization should not be in the store')
     clear()
 
-    user = await store.create('user', store.createRecord('user', JSData.utils.copy(userProps)), { pass: ['comment', 'profile'] })
+    user = await store.create('user', store.createRecord('user', JSData.utils.plainCopy(userProps)), { pass: ['comments', 'profile'] })
     assert(store.is('user', user), 'user should be a user record')
     assert.strictEqual(store.get('user', user.id), user, 'user should be in the store')
     assert(store.is('comment', user.comments[0]), 'user.comments[0] should be a comment record')
@@ -334,7 +367,7 @@ describe('Mapper#create', function () {
     assert.deepEqual(store.getAll('organization'), [], 'organization should not be in the store')
     clear()
 
-    user = await store.create('user', store.createRecord('user', JSData.utils.copy(userProps)), { pass: ['comment', 'profile', 'organization'] })
+    user = await store.create('user', store.createRecord('user', JSData.utils.plainCopy(userProps)), { pass: ['comments', 'profile', 'organization'] })
     assert(store.is('user', user), 'user should be a user record')
     assert.strictEqual(store.get('user', user.id), user, 'user should be in the store')
     assert(store.is('comment', user.comments[0]), 'user.comments[0] should be a comment record')
@@ -347,43 +380,58 @@ describe('Mapper#create', function () {
     clear()
 
     // when props are NOT a record
-    user = await store.create('user', JSData.utils.copy(userProps), { pass: [] })
+    user = await store.create('user', JSData.utils.plainCopy(userProps), { pass: [] })
     assert(store.is('user', user), 'user should be a user record')
     assert.strictEqual(store.get('user', user.id), user, 'user should be in the store')
-    assert.deepEqual(user.comments, [], 'user.comments should be an empty array')
+    assert.objectsEqual(user.comments, [
+      {
+        content: 'foo',
+        userId: user.id
+      }
+    ], 'user.comments should be an empty array')
     assert.deepEqual(store.getAll('comment'), [], 'comments should not be in the store')
-    assert(!user.profile, 'user.profile should be undefined')
+    assert.objectsEqual(user.profile, {
+      email: 'john@email.com'
+    }, 'user.profile should be a profile')
     assert.deepEqual(store.getAll('profile'), [], 'profile should not be in the store')
-    assert(!user.organization, 'user.organization should be undefined')
+    assert.objectsEqual(user.organization, {
+      name: 'Company Inc'
+    }, 'user.organization should be an organization')
     assert(!user.organizationId, 'user.organizationId should be undefined')
     assert.deepEqual(store.getAll('organization'), [], 'organization should not be in the store')
     clear()
 
-    user = await store.create('user', JSData.utils.copy(userProps), { pass: ['comment'] })
+    user = await store.create('user', JSData.utils.plainCopy(userProps), { pass: ['comments'] })
     assert(store.is('user', user), 'user should be a user record')
     assert.strictEqual(store.get('user', user.id), user, 'user should be in the store')
     assert(store.is('comment', user.comments[0]), 'user.comments[0] should be a comment record')
     assert.objectsEqual(store.getAll('comment'), user.comments, 'comments should be in the store')
-    assert(!user.profile, 'user.profile should be undefined')
+    assert.objectsEqual(user.profile, {
+      email: 'john@email.com'
+    }, 'user.profile should be a profile')
     assert.deepEqual(store.getAll('profile'), [], 'profile should not be in the store')
-    assert(!user.organization, 'user.organization should be undefined')
+    assert.objectsEqual(user.organization, {
+      name: 'Company Inc'
+    }, 'user.organization should be an organization')
     assert(!user.organizationId, 'user.organizationId should be undefined')
     assert.deepEqual(store.getAll('organization'), [], 'organization should not be in the store')
     clear()
 
-    user = await store.create('user', JSData.utils.copy(userProps), { pass: ['comment', 'profile'] })
+    user = await store.create('user', JSData.utils.plainCopy(userProps), { pass: ['comments', 'profile'] })
     assert(store.is('user', user), 'user should be a user record')
     assert.strictEqual(store.get('user', user.id), user, 'user should be in the store')
     assert(store.is('comment', user.comments[0]), 'user.comments[0] should be a comment record')
     assert.objectsEqual(store.getAll('comment'), user.comments, 'comments should be in the store')
     assert(store.is('profile', user.profile), 'user.profile should be a profile record')
     assert.objectsEqual(store.getAll('profile'), [user.profile], 'profile should be in the store')
-    assert(!user.organization, 'user.organization should be undefined')
+    assert.objectsEqual(user.organization, {
+      name: 'Company Inc'
+    }, 'user.organization should be an organization')
     assert(!user.organizationId, 'user.organizationId should be undefined')
     assert.deepEqual(store.getAll('organization'), [], 'organization should not be in the store')
     clear()
 
-    user = await store.create('user', JSData.utils.copy(userProps), { pass: ['comment', 'profile', 'organization'] })
+    user = await store.create('user', JSData.utils.plainCopy(userProps), { pass: ['comments', 'profile', 'organization'] })
     assert(store.is('user', user), 'user should be a user record')
     assert.strictEqual(store.get('user', user.id), user, 'user should be in the store')
     assert(store.is('comment', user.comments[0]), 'user.comments[0] should be a comment record')
@@ -458,31 +506,40 @@ describe('Mapper#create', function () {
     }
 
     // when props are a Record
-    let user = await store.create('user', store.createRecord('user', JSData.utils.copy(userProps)), { pass: [] })
+    let user = await store.create('user', store.createRecord('user', JSData.utils.plainCopy(userProps)), { pass: [] })
     assert(store.is('user', user), 'user should be a user record')
     assert.strictEqual(store.get('user', user.id), user, 'user should be in the store')
-    assert.deepEqual(user.comments, [], 'user.comments should be an empty array')
+    assert.objectsEqual(user.comments, [
+      {
+        content: 'foo',
+        userId: user.id
+      }
+    ], 'user.comments should be an array')
     assert.deepEqual(store.getAll('comment'), [], 'comments should not be in the store')
-    assert(!user.profile, 'user.profile should be undefined')
+    assert.objectsEqual(user.profile, {
+      email: 'john@email.com'
+    }, 'user.profile should be a profile')
     assert.deepEqual(store.getAll('profile'), [], 'profile should not be in the store')
     assert(!user.organization, 'user.organization should be undefined')
     assert(!user.organizationId, 'user.organizationId should be undefined')
     assert.deepEqual(store.getAll('organization'), [], 'organization should not be in the store')
     clear()
 
-    user = await store.create('user', store.createRecord('user', JSData.utils.copy(userProps)), { pass: ['comment'] })
+    user = await store.create('user', store.createRecord('user', JSData.utils.plainCopy(userProps)), { pass: ['comments'] })
     assert(store.is('user', user), 'user should be a user record')
     assert.strictEqual(store.get('user', user.id), user, 'user should be in the store')
     assert(store.is('comment', user.comments[0]), 'user.comments[0] should be a comment record')
     assert.objectsEqual(store.getAll('comment'), user.comments, 'comments should be in the store')
-    assert(!user.profile, 'user.profile should be undefined')
+    assert.objectsEqual(user.profile, {
+      email: 'john@email.com'
+    }, 'user.profile should be a profile')
     assert.deepEqual(store.getAll('profile'), [], 'profile should not be in the store')
     assert(!user.organization, 'user.organization should be undefined')
     assert(!user.organizationId, 'user.organizationId should be undefined')
     assert.deepEqual(store.getAll('organization'), [], 'organization should not be in the store')
     clear()
 
-    user = await store.create('user', store.createRecord('user', JSData.utils.copy(userProps)), { with: ['comment'], pass: ['profile'] })
+    user = await store.create('user', store.createRecord('user', JSData.utils.plainCopy(userProps)), { with: ['comments'], pass: ['profile'] })
     assert(store.is('user', user), 'user should be a user record')
     assert.strictEqual(store.get('user', user.id), user, 'user should be in the store')
     assert(store.is('comment', user.comments[0]), 'user.comments[0] should be a comment record')
@@ -494,7 +551,7 @@ describe('Mapper#create', function () {
     assert.deepEqual(store.getAll('organization'), [], 'organization should not be in the store')
     clear()
 
-    user = await store.create('user', store.createRecord('user', JSData.utils.copy(userProps)), { with: ['comment', 'profile'], pass: ['organization'] })
+    user = await store.create('user', store.createRecord('user', JSData.utils.plainCopy(userProps)), { with: ['comments', 'profile'], pass: ['organization'] })
     assert(store.is('user', user), 'user should be a user record')
     assert.strictEqual(store.get('user', user.id), user, 'user should be in the store')
     assert(store.is('comment', user.comments[0]), 'user.comments[0] should be a comment record')
@@ -507,43 +564,58 @@ describe('Mapper#create', function () {
     clear()
 
     // when props are NOT a record
-    user = await store.create('user', JSData.utils.copy(userProps), { pass: [] })
+    user = await store.create('user', JSData.utils.plainCopy(userProps), { pass: [] })
     assert(store.is('user', user), 'user should be a user record')
     assert.strictEqual(store.get('user', user.id), user, 'user should be in the store')
-    assert.deepEqual(user.comments, [], 'user.comments should be an empty array')
+    assert.objectsEqual(user.comments, [
+      {
+        content: 'foo',
+        userId: user.id
+      }
+    ], 'user.comments should be an array')
     assert.deepEqual(store.getAll('comment'), [], 'comments should not be in the store')
-    assert(!user.profile, 'user.profile should be undefined')
+    assert.objectsEqual(user.profile, {
+      email: 'john@email.com'
+    }, 'user.profile should be a profile')
     assert.deepEqual(store.getAll('profile'), [], 'profile should not be in the store')
-    assert(!user.organization, 'user.organization should be undefined')
+    assert.objectsEqual(user.organization, {
+      name: 'Company Inc'
+    }, 'user.organization should be an organization')
     assert(!user.organizationId, 'user.organizationId should be undefined')
     assert.deepEqual(store.getAll('organization'), [], 'organization should not be in the store')
     clear()
 
-    user = await store.create('user', JSData.utils.copy(userProps), { pass: ['comment'] })
+    user = await store.create('user', JSData.utils.plainCopy(userProps), { pass: ['comments'] })
     assert(store.is('user', user), 'user should be a user record')
     assert.strictEqual(store.get('user', user.id), user, 'user should be in the store')
     assert(store.is('comment', user.comments[0]), 'user.comments[0] should be a comment record')
     assert.objectsEqual(store.getAll('comment'), user.comments, 'comments should be in the store')
-    assert(!user.profile, 'user.profile should be undefined')
+    assert.objectsEqual(user.profile, {
+      email: 'john@email.com'
+    }, 'user.profile should be a profile')
     assert.deepEqual(store.getAll('profile'), [], 'profile should not be in the store')
-    assert(!user.organization, 'user.organization should be undefined')
+    assert.objectsEqual(user.organization, {
+      name: 'Company Inc'
+    }, 'user.organization should be an organization')
     assert(!user.organizationId, 'user.organizationId should be undefined')
     assert.deepEqual(store.getAll('organization'), [], 'organization should not be in the store')
     clear()
 
-    user = await store.create('user', JSData.utils.copy(userProps), { with: ['comment'], pass: ['profile'] })
+    user = await store.create('user', JSData.utils.plainCopy(userProps), { with: ['comments'], pass: ['profile'] })
     assert(store.is('user', user), 'user should be a user record')
     assert.strictEqual(store.get('user', user.id), user, 'user should be in the store')
     assert(store.is('comment', user.comments[0]), 'user.comments[0] should be a comment record')
     assert.objectsEqual(store.getAll('comment'), user.comments, 'comments should be in the store')
     assert(store.is('profile', user.profile), 'user.profile should be a profile record')
     assert.objectsEqual(store.getAll('profile'), [user.profile], 'profile should be in the store')
-    assert(!user.organization, 'user.organization should be undefined')
+    assert.objectsEqual(user.organization, {
+      name: 'Company Inc'
+    }, 'user.organization should be an organization')
     assert(!user.organizationId, 'user.organizationId should be undefined')
     assert.deepEqual(store.getAll('organization'), [], 'organization should not be in the store')
     clear()
 
-    user = await store.create('user', JSData.utils.copy(userProps), { with: ['comment', 'profile'], pass: ['organization'] })
+    user = await store.create('user', JSData.utils.plainCopy(userProps), { with: ['comments', 'profile'], pass: ['organization'] })
     assert(store.is('user', user), 'user should be a user record')
     assert.strictEqual(store.get('user', user.id), user, 'user should be in the store')
     assert(store.is('comment', user.comments[0]), 'user.comments[0] should be a comment record')
@@ -556,7 +628,7 @@ describe('Mapper#create', function () {
     clear()
 
     assert.equal(createCalledCount.user, 8)
-    assert.equal(createCalledCount.comment, 6)
+    assert.equal(createCalledCount.comment, 4)
     assert.equal(createCalledCount.profile, 2)
     assert(!createCalledCount.organization)
   })